requirements.txt
Arpeggio==2.0.2
attrs==23.1.0
dill==0.3.7
grpcio==1.59.2
parver==0.5
protobuf==4.25.0
pulumi==3.91.1
pulumi-fortios==0.0.9
PyYAML==6.0.1
semver==2.13.0
six==1.16.0
